requirements.txt
# unit tests
pytest

# OpenAI API for LM Studio access
openai

scipy
audioop-lts
sounddevice
soundfile
pydub

# TTS
piper-tts
onnxruntime-gpu
# run manually: python -m piper.download_voices en_US-kristin-medium

# speaker separation
speechbrain

# STT
faster-whisper

numpy>=1.24.0
--extra-index-url https://download.pytorch.org/whl/cu126
torch>=2.0.0
torchaudio>=2.0.0,<2.9.0
accelerate
transformers>=4.30.0
tokenizers>=0.13.0

# visualizations
matplotlib

# VAD for speech detection
webrtcvad

# AEC (Acoustic Echo Cancellation)
aec-audio-processing